fix(type declaration): define the `OutputData` type in the `postinstall-debug-package/types.ts` file instead of the `postinstall-debug-package/postinstall.ts` file
The `postinstall-debug-package/postinstall.ts` file uses type declarations that cannot be built for Node.js 16.
